sort imports, add rating

+++ api_yamdb/api/serializers.py
@@ -1,6 +1,7 @@
+from django.core.validators import MaxValueValidator, MinValueValidator
 from rest_framework import serializers
-from posts.models import Category, Genre, Title, Review, Comment
-from django.core.validators import MinValueValidator, MaxValueValidator
+
+from posts.models import Category, Comment, Genre, Review, Title
 
 
 class CategorySerializer(serializers.ModelSerializer):
@@ -20,7 +21,7 @@ class GenreSerializer(serializers.ModelSerializer):
 class TitleSerializer(serializers.ModelSerializer):
     genre = GenreSerializer(many=True, read_only=True)
     category = CategorySerializer(read_only=True)
-    rating = serializers.IntegerField(read_only=True)
+    rating = serializers.FloatField(read_only=True)
 
     class Meta:
         fields = '__all__'
